Add Google Analytics support
