+++ client/src/App.js
@@ -1,8 +1,9 @@
-import React, { useState, useEffect } from "react";
-import { BrowserRouter as Router, Route, Redirect } from "react-router-dom";
+import React from "react";
+import PropTypes from "prop-types";
+import { Route, Redirect } from "react-router-dom";
+import { useTracking } from "./hooks/useTracking";
 import { createUseStyles } from "react-jss";
 import { withToastProvider } from "./contexts/Toast";
-import { UserContext } from "./components/user-context";
 import TdmCalculationContainer from "./components/TdmCalculationContainer";
 import ProjectsPage from "./components/Projects/ProjectsPage";
 import Header from "./components/Header";
@@ -21,7 +22,6 @@ import ResetPassword from "./components/Authorization/ResetPassword";
 import ResetPasswordRequest from "./components/Authorization/ResetPasswordRequest";
 import "./styles/App.scss";
 import PublicComment from "./components/PublicComment/PublicCommentPage";
-import NavConfirmModal from "./components/NavConfirmModal";
 
 const useStyles = createUseStyles({
   root: {
@@ -31,131 +31,97 @@ const useStyles = createUseStyles({
   }
 });
 
-const App = () => {
+const App = ({ account, setLoggedInAccount, hasConfirmedTransition }) => {
   const classes = useStyles();
-  const [account, setAccount] = useState({});
-  const [confirmTransition, setConfirmTransition] = useState(null);
-  const [hasConfirmedTransition, setHasConfirmedTransition] = useState(true);
-  const [isOpenNavConfirmModal, setIsOpenNavConfirmModal] = useState(false);
-
-  const setLoggedInAccount = loggedInUser => {
-    setAccount(loggedInUser);
-    localStorage.setItem("currentUser", JSON.stringify(loggedInUser));
-  };
-
-  useEffect(() => {
-    const currentUser = localStorage.getItem("currentUser");
-    if (currentUser) {
-      try {
-        const parsedAccount = JSON.parse(currentUser);
-        setAccount(parsedAccount);
-      } catch (err) {
-        // TODO: replace with production error logging.
-        console.error(
-          "Unable to parse current user from local storage.",
-          currentUser
-        );
-      }
-    } else {
-      setLoggedInAccount({});
-    }
-  }, []);
-
-  const getUserConfirmation = (_message, defaultConfirmCallback) => {
-    setHasConfirmedTransition(false);
-    setConfirmTransition(() => ({
-      defaultConfirmCallback: defaultConfirmCallback,
-      setHasConfirmedTransition: setHasConfirmedTransition
-    }));
-    setIsOpenNavConfirmModal(!isOpenNavConfirmModal);
-  };
+  useTracking("G-MW23VES3G6");
 
   return (
     <React.Fragment>
-      <UserContext.Provider value={account}>
-        <Router getUserConfirmation={getUserConfirmation}>
-          <NavConfirmModal
-            confirmTransition={confirmTransition}
-            isOpenNavConfirmModal={isOpenNavConfirmModal}
-            setIsOpenNavConfirmModal={setIsOpenNavConfirmModal}
-          />
-          <Header account={account} />
-          <div className={classes.root}>
-            <Route
-              exact
-              path="/"
-              render={() =>
-                account.email ? (
-                  <Redirect to="/create-project" />
-                ) : (
-                  <Login setLoggedInAccount={setLoggedInAccount} />
-                )
-              }
+      <Header account={account} />
+      <div className={classes.root}>
+        <Route
+          exact
+          path="/"
+          render={() =>
+            account.email ? (
+              <Redirect to="/create-project" />
+            ) : (
+              <Login setLoggedInAccount={setLoggedInAccount} />
+            )
+          }
+        />
+        <Route exact path="/create-project">
+          <Redirect to="/calculation" />
+        </Route>
+        <Route
+          path="/calculation/:page?/:projectId?"
+          render={() => (
+            <TdmCalculationContainer
+              account={account}
+              hasConfirmedNavTransition={hasConfirmedTransition}
+              setLoggedInAccount={setLoggedInAccount}
             />
-            <Route exact path="/create-project">
-              <Redirect to="/calculation" />
-            </Route>
-            <Route
-              path="/calculation/:page?/:projectId?"
-              render={() => (
-                <TdmCalculationContainer
-                  account={account}
-                  hasConfirmedNavTransition={hasConfirmedTransition}
-                  setLoggedInAccount={setLoggedInAccount}
-                />
-              )}
-            />
-            <Route
-              path="/projects"
-              render={() => <ProjectsPage account={account} />}
-            />
-            <Route path="/about" component={About} />
-            <Route path="/termsandconditions" component={TermsAndConditions} />
-            <Route path="/privacypolicy" component={PrivacyPolicy} />
-            <Route path="/register/:email?" component={Register} />
-            <Route path="/confirm/:token" component={ConfirmEmail} />
-            <Route
-              path="/login/:email?"
-              render={() =>
-                account.email ? (
-                  <Redirect to="/create-project" />
-                ) : (
-                  <Login setLoggedInAccount={setLoggedInAccount} />
-                )
-              }
-            />
-            <Route
-              path="/logout/:email?"
-              render={routeProps => {
-                setLoggedInAccount({});
-                return (
-                  <Redirect
-                    to={
-                      routeProps.match.params["email"]
-                        ? `/login/${routeProps.match.params["email"]}`
-                        : "/login"
-                    }
-                  />
-                );
-              }}
-            />
-            <Route path="/forgotpassword" component={ResetPasswordRequest} />
-            <Route path="/resetPassword/:token" component={ResetPassword} />
-            <Route path="/contactus" component={ContactUs} />
-            {account && account.isAdmin ? (
-              <Route path="/admin" render={() => <Admin account={account} />} />
-            ) : null}
-            {account && account.isSecurityAdmin ? (
-              <Route path="/roles" render={() => <Roles />} />
-            ) : null}
-            <Route path="/faqs" component={FaqView} />
-            <Route path="/publiccomment" component={PublicComment} />
-          </div>
-          <Footer />
-        </Router>
-      </UserContext.Provider>
+          )}
+        />
+        <Route
+          path="/projects"
+          render={() => <ProjectsPage account={account} />}
+        />
+        <Route path="/about" component={About} />
+        <Route path="/termsandconditions" component={TermsAndConditions} />
+        <Route path="/privacypolicy" component={PrivacyPolicy} />
+        <Route path="/register/:email?" component={Register} />
+        <Route path="/confirm/:token" component={ConfirmEmail} />
+        <Route
+          path="/login/:email?"
+          render={() =>
+            account.email ? (
+              <Redirect to="/create-project" />
+            ) : (
+              <Login setLoggedInAccount={setLoggedInAccount} />
+            )
+          }
+        />
+        <Route
+          path="/logout/:email?"
+          render={routeProps => {
+            setLoggedInAccount({});
+            return (
+              <Redirect
+                to={
+                  routeProps.match.params["email"]
+                    ? `/login/${routeProps.match.params["email"]}`
+                    : "/login"
+                }
+              />
+            );
+          }}
+        />
+        <Route path="/forgotpassword" component={ResetPasswordRequest} />
+        <Route path="/resetPassword/:token" component={ResetPassword} />
+        <Route path="/contactus" component={ContactUs} />
+        {account && account.isAdmin ? (
+          <Route path="/admin" render={() => <Admin account={account} />} />
+        ) : null}
+        {account && account.isSecurityAdmin ? (
+          <Route path="/roles" render={() => <Roles />} />
+        ) : null}
+        <Route path="/faqs" component={FaqView} />
+        <Route path="/publiccomment" component={PublicComment} />
+      </div>
+      <Footer />
     </React.Fragment>
   );
+};
+
+App.propTypes = {
+  account: PropTypes.shape({
+    email: PropTypes.string,
+    isAdmin: PropTypes.bool,
+    isSecurityAdmin: PropTypes.bool
+  }),
+  setLoggedInAccount: PropTypes.func,
+  hasConfirmedTransition: PropTypes.func
 };
 
 export default withToastProvider(App);
