+++ client/src/index.js
@@ -1,13 +1,13 @@
 import React from "react";
 import ReactDom from "react-dom";
-import App from "./App";
+import AppWrapper from "./AppWrapper";
 import { ThemeProvider } from "react-jss";
 import { jssTheme } from "./styles/theme";
 
 try {
   ReactDom.render(
     <ThemeProvider theme={jssTheme}>
-      <App />
+      <AppWrapper />
     </ThemeProvider>,
     document.querySelector("#root")
   );
